generic can edit func
