+++ internal/ent/hooks/programmembers.go
@@ -13,7 +13,7 @@ import (
 func HookProgramMembers() ent.Hook {
 	return hook.On(func(next ent.Mutator) ent.Mutator {
 		return hook.ProgramMembershipFunc(func(ctx context.Context, m *generated.ProgramMembershipMutation) (generated.Value, error) {
-			// if userID is on the mutation then we need dto check if the user is a member of the organization
+			// if userID is on the mutation then we need to check if the user is a member of the organization
 			userID, ok := m.UserID()
 			if !ok {
 				return next.Mutate(ctx, m)
@@ -26,7 +26,7 @@ func HookProgramMembers() ent.Hook {
 
 			program, err := m.Client().Program.Get(ctx, programID)
 			if err != nil {
-				// group not found, let the default validation handle it
+				// program not found, let the default validation handle it
 				return next.Mutate(ctx, m)
 			}
