+++ internal/ent/schema/program.go
@@ -1,13 +1,18 @@
 package schema
 
 import (
+	"context"
+
 	"entgo.io/contrib/entgql"
 	"entgo.io/ent"
 	"entgo.io/ent/schema"
 	"entgo.io/ent/schema/edge"
 	"entgo.io/ent/schema/field"
 
+	"github.com/theopenlane/core/internal/ent/generated"
+	"github.com/theopenlane/core/internal/ent/generated/privacy"
 	"github.com/theopenlane/core/internal/ent/mixin"
+	"github.com/theopenlane/core/internal/ent/privacy/rule"
 	"github.com/theopenlane/core/pkg/enums"
 	"github.com/theopenlane/entx"
 	emixin "github.com/theopenlane/entx/mixin"
@@ -132,4 +137,35 @@ func (Program) Hooks() []ent.Hook {
 // Interceptors of the Program
 func (Program) Interceptors() []ent.Interceptor {
 	return []ent.Interceptor{}
+}
+
+// Policy of the program
+func (Program) Policy() ent.Policy {
+	return privacy.Policy{
+		Mutation: privacy.MutationPolicy{
+			privacy.OnMutationOperation(
+				rule.CanCreateObjectsInOrg(),
+				ent.OpCreate,
+			),
+			privacy.OnMutationOperation(
+				privacy.ProgramMutationRuleFunc(func(ctx context.Context, m *generated.ProgramMutation) error {
+					return m.CheckAccessForEdit(ctx)
+				}),
+				ent.OpUpdate|ent.OpUpdateOne|ent.OpUpdate,
+			),
+			privacy.OnMutationOperation(
+				privacy.ProgramMutationRuleFunc(func(ctx context.Context, m *generated.ProgramMutation) error {
+					return m.CheckAccessForDelete(ctx)
+				}),
+				ent.OpDelete|ent.OpDeleteOne,
+			),
+			privacy.AlwaysDenyRule(),
+		},
+		Query: privacy.QueryPolicy{
+			privacy.ProgramQueryRuleFunc(func(ctx context.Context, q *generated.ProgramQuery) error {
+				return q.CheckAccess(ctx)
+			}),
+			privacy.AlwaysDenyRule(),
+		},
+	}
 }
